+++ src/main/scala/io/smartdatalake/workflow/ActionDAGRunState.scala
@@ -70,97 +70,28 @@ private[smartdatalake]object ActionDAGRunState {
   ))
 }
 
-private[smartdatalake] case class ActionDAGRunStateStore(statePath: String, appName: String) extends SmartDataLakeLogger {
-
-  private val hadoopStatePath = HdfsUtil.addHadoopDefaultSchemaAuthority(new Path(statePath))
-  private val currentStatePath = new Path(hadoopStatePath, "current")
-  private val succeededStatePath = new Path(hadoopStatePath, "succeeded")
-  implicit private val filesystem: FileSystem = HdfsUtil.getHadoopFs(hadoopStatePath)
-  if (!filesystem.exists(hadoopStatePath)) filesystem.mkdirs(hadoopStatePath)
-  filesystem.setWriteChecksum(false) // disable writing CRC files
+private[smartdatalake] trait ActionDAGRunStateStore[A <: StateId] extends SmartDataLakeLogger {
 
   /**
-   * Save state to file
+   * Save State
    */
-  def saveState(state: ActionDAGRunState): Unit = synchronized {
-    val path = if (state.isSucceeded) succeededStatePath else currentStatePath
-    // write state file
-    val json = state.toJson
-    val fileName = s"$appName${ActionDAGRunStateStore.fileNamePartSeparator}${state.runId}${ActionDAGRunStateStore.fileNamePartSeparator}${state.attemptId}.json"
-    val file = new Path(path, fileName)
-    val os = filesystem.create(file, true) // overwrite if exists
-    os.write(json.getBytes("UTF-8"))
-    os.close()
-    logger.info(s"updated state into ${file.toUri}")
-    // if succeeded:
-    // - delete temporary state file from current directory
-    // - move previous failed attempt files from current to succeeded directory
-    if (state.isSucceeded) {
-      filesystem.delete(new Path(currentStatePath, fileName), /*recursive*/ false)
-      getFiles(Some(currentStatePath))
-        .filter( stateFile => stateFile.runId==state.runId && stateFile.attemptId<state.attemptId)
-        .foreach { stateFile =>
-          logger.info(s"renamed ${stateFile.path}")
-          FileUtil.copy(filesystem, stateFile.path, filesystem, succeededStatePath, /*deleteSource*/ true, filesystem.getConf)
-        }
-    }
-  }
+  def saveState(state: ActionDAGRunState): Unit
 
   /**
    * Get latest state
    * @param runId optional runId to search for latest state
    */
-  def getLatestState(runId: Option[Int] = None): StateFile = {
-    val latestStateFile = getFiles()
-      .filter(x => runId.isEmpty || runId.contains(x.runId))
-      .sortBy(_.getSortAttrs).lastOption
-    require(latestStateFile.nonEmpty, s"No state file for application $appName and runId ${runId.getOrElse("latest")} found.")
-    latestStateFile.get
-  }
+  def getLatestState(runId: Option[Int] = None): A
 
   /**
    * Get latest runId
    */
-  def getLatestRunId: Option[Int] = {
-    val latestStateFile = getFiles()
-      .sortBy(_.getSortAttrs).lastOption
-    latestStateFile.map(_.runId)
-  }
-
-  /**
-   * Search state directory for state files of this app
-   */
-  private def getFiles(path: Option[Path] = None): Seq[StateFile] = {
-    val filenameMatcher = s"([^_]+)\\${ActionDAGRunStateStore.fileNamePartSeparator}([0-9]+)\\${ActionDAGRunStateStore.fileNamePartSeparator}([0-9]+)\\.json".r
-    val pathFilter = new PathFilter {
-      override def accept(path: Path): Boolean = path.getName.startsWith(appName + ActionDAGRunStateStore.fileNamePartSeparator)
-    }
-    val searchPath = path.getOrElse( new Path(hadoopStatePath, "*"))
-    filesystem.globStatus(new Path(searchPath, "*.json"), pathFilter )
-      .filter( x => x.isFile)
-      .flatMap( x => x.getPath.getName match {
-        case filenameMatcher(appName, runId, attemptId) =>
-          Some(StateFile(x.getPath, appName, runId.toInt, attemptId.toInt))
-        case _ => None
-      })
-      .filter(_.appName == this.appName)
-  }
+  def getLatestRunId: Option[Int]
 
   /**
    * recover previous run state
    */
-  def recoverRunState(stateFile: Path): ActionDAGRunState = {
-    require(filesystem.isFile(stateFile), s"Cannot recover previous run state. ${stateFile.toUri} doesn't exists or is not a file.")
-    val is = filesystem.open(stateFile)
-    val json = scala.io.Source.fromInputStream(is)(Codec.UTF8).mkString
-    ActionDAGRunState.fromJson(json)
-  }
-
-  case class StateFile(path: Path, appName: String, runId: Int, attemptId: Int) {
-    def getSortAttrs: (Int, Int) = (runId, attemptId)
-  }
+  def recoverRunState(stateId: A): ActionDAGRunState
 }
 
-private[smartdatalake] object ActionDAGRunStateStore {
-  val fileNamePartSeparator = "."
-}
+private[smartdatalake] trait StateId
